fix: renamed universal_api to dial_api
