+++ aidial_adapter_bedrock/app.py
@@ -7,13 +7,10 @@ from fastapi import Request
 from fastapi.responses import JSONResponse
 
 from aidial_adapter_bedrock.chat_completion import BedrockChatCompletion
+from aidial_adapter_bedrock.dial_api.response import ModelObject, ModelsResponse
 from aidial_adapter_bedrock.llm.bedrock_models import BedrockDeployment
 from aidial_adapter_bedrock.llm.model_listing import get_bedrock_models
 from aidial_adapter_bedrock.server.exceptions import dial_exception_decorator
-from aidial_adapter_bedrock.universal_api.response import (
-    ModelObject,
-    ModelsResponse,
-)
 from aidial_adapter_bedrock.utils.env import get_env
 from aidial_adapter_bedrock.utils.log_config import LogConfig
 from aidial_adapter_bedrock.utils.log_config import app_logger as log
