+++ client/client_bedrock.py
@@ -3,10 +3,10 @@ from typing import List
 
 from aidial_sdk.chat_completion import Message, Role
 
+from aidial_adapter_bedrock.dial_api.request import ModelParameters
 from aidial_adapter_bedrock.llm.bedrock_models import BedrockDeployment
 from aidial_adapter_bedrock.llm.consumer import CollectConsumer
 from aidial_adapter_bedrock.llm.model.adapter import get_bedrock_adapter
-from aidial_adapter_bedrock.universal_api.request import ModelParameters
 from aidial_adapter_bedrock.utils.env import get_env
 from aidial_adapter_bedrock.utils.printing import print_ai, print_info
 from client.conf import MAX_CHAT_TURN, MAX_INPUT_CHARS
